Remove public pages, move support to private
