+++ app/(auth)/login/page.tsx
@@ -1,9 +1,8 @@
+import AuthFormSelector from '@/components/auth-form-selector'
 import { Icons } from '@/components/icons'
-import { buttonVariants } from '@/components/ui/button'
-import { cn } from '@/lib/utils'
+import { MainNav } from '@/components/main-nav'
 import { Metadata } from 'next'
 import Link from 'next/link'
-import AuthFormSelector from '@/components/auth-form-selector'
 
 export const metadata: Metadata = {
   title: 'Login',
@@ -13,40 +12,31 @@ export const metadata: Metadata = {
 export default function LoginPage() {
   return (
     <div className="container flex h-screen w-screen flex-col items-center justify-center">
-      <Link
-        href="/"
-        className={cn(
-          buttonVariants({ variant: 'ghost' }),
-          'absolute left-4 top-4 md:left-8 md:top-8'
-        )}
-      >
-        <>
-          <Icons.chevronLeft className="mr-2 h-4 w-4 text-white" />
-          Back
-        </>
-      </Link>
+      <div className="absolute left-4 top-4 md:left-8 md:top-8">
+        <MainNav />
+      </div>
       <div className="bg-dark text-white lg:p-8 ">
         <div className="m-auto flex w-full flex-col justify-center space-y-6 ">
           <div className="flex flex-col space-y-2 text-center  ">
-          <Icons.logo className="mx-auto h-6 w-6" />
-          <h1 className="text-2xl font-semibold tracking-tight">
-            Welcome back
-          </h1>
-          <p className="text-sm text-muted-foreground">
-            Enter your email to sign in to your account
+            <Icons.logo className="mx-auto h-6 w-6" />
+            <h1 className="text-2xl font-semibold tracking-tight">
+              Welcome back
+            </h1>
+            <p className="text-sm text-muted-foreground">
+              Enter your email to sign in to your account
+            </p>
+          </div>
+          <AuthFormSelector />
+          <p className="px-8 text-center text-sm text-muted-foreground">
+            <Link
+              href="/signUp"
+              className="underline underline-offset-4 hover:text-accent-1"
+            >
+              Don&apos;t have an account? Sign Up
+            </Link>
           </p>
         </div>
-        <AuthFormSelector />
-        <p className="px-8 text-center text-sm text-muted-foreground">
-          <Link
-            href="/signUp"
-            className="underline underline-offset-4 hover:text-accent-1"
-          >
-            Don&apos;t have an account? Sign Up
-          </Link>
-        </p>
       </div>
-    </div>
     </div>
   )
 }
