--- app/(public)/blog/[slug]/head.tsx
@@ -1,31 +1,0 @@
-import BlogMeta from '@/components/blog/BlogMeta'
-import * as demo from 'lib/demo.data'
-import { getPostBySlug, getSettings } from 'lib/sanity.client'
-import { urlForImage } from 'lib/sanity.image'
-
-export default async function SlugHead({
-  params,
-}: {
-  params: { slug: string }
-}) {
-  const [{ title = demo.title }, post] = await Promise.all([
-    getSettings(),
-    getPostBySlug(params.slug),
-  ])
-  return (
-    <>
-      <title>{post.title ? `${post.title} | ${title}` : title}</title>
-      <BlogMeta />
-      {post.coverImage?.asset?._ref && (
-        <meta
-          property="og:image"
-          content={urlForImage(post.coverImage)
-            .width(1200)
-            .height(627)
-            .fit('crop')
-            .url()}
-        />
-      )}
-    </>
-  )
-}
