+++ components/main-nav.tsx
@@ -20,7 +20,10 @@ export function MainNav({ items, children }: MainNavProps) {
 
   return (
     <div className="flex gap-6 md:gap-10">
-      <Link href="/" className="hidden items-center space-x-2 md:flex">
+      <Link
+        href="https://donorsync.org"
+        className="hidden items-center space-x-2 md:flex"
+      >
         <Icons.logo />
         <span className="hidden font-bold sm:inline-block">
           {siteConfig.name}
