+++ config/dashboard.ts
@@ -1,34 +1,7 @@
 import { DashboardConfig } from 'types'
 
 export const dashboardConfig: DashboardConfig = {
-  mainNav: [
-    {
-      title: 'Features',
-      href: '/features',
-      // disabled: true,
-    },
-    {
-      title: 'Pricing',
-      href: '/pricing',
-    },
-    {
-      title: 'Blog',
-      href: '/blog',
-    },
-    {
-      title: 'Documentation',
-      href: '/docs',
-    },
-    {
-      title: 'Support',
-      href: '/support',
-    },
-    {
-      title: 'Contact',
-      href: '/contact',
-      disabled: true,
-    },
-  ],
+  mainNav: [],
   sidebarNav: [
     {
       title: 'Dashboard',
@@ -54,6 +27,10 @@ export const dashboardConfig: DashboardConfig = {
       title: 'Settings',
       href: '/dashboard/settings',
       icon: 'settings',
+    },
+    {
+      title: 'Support',
+      href: '/support',
     },
   ],
 }
