--- config/marketing.ts
@@ -1,32 +1,0 @@
-import { MarketingConfig } from 'types'
-
-export const marketingConfig: MarketingConfig = {
-  mainNav: [
-    {
-      title: 'Features',
-      href: '/features',
-      // disabled: true,
-    },
-    {
-      title: 'Pricing',
-      href: '/pricing',
-    },
-    {
-      title: 'Blog',
-      href: '/blog',
-    },
-    {
-      title: 'Documentation',
-      href: '/docs',
-    },
-    {
-      title: 'Support',
-      href: '/support',
-    },
-    {
-      title: 'Contact',
-      href: '/contact',
-      disabled: true,
-    },
-  ],
-}
